chore: code quality cleanup — fix imports, stale comments, type placement
