+++ components/sports/session-views/custom-ordered-view.tsx
@@ -1,6 +1,6 @@
 "use client";
 
-import { Fragment, useState } from "react";
+import { useState } from "react";
 import {
     Table,
     TableBody,
