+++ components/sports/session-views/interfaces.ts
@@ -1,20 +1,6 @@
 import type { ComponentType } from "react";
 import type { SignupRow } from "@/components/sports/session-signups-table";
 
-/** Shape of each view instance stored in the session's views JSONB array. */
-export interface StoredViewInstance {
-    /** Stable numeric identifier. Also determines display order (ascending). */
-    id: number;
-    /** Registry key (e.g. "customOrderedView") — determines which component renders it. */
-    type: string;
-    /** Admin-given display name (e.g. "Batting Order"). */
-    label: string;
-    /** View-specific payload. */
-    data: unknown;
-    /** Whether this view is visible in the toggle. Defaults to true if omitted. */
-    enabled?: boolean;
-}
-
 /** Props passed to every session view component. */
 export interface SessionViewProps {
     signups: SignupRow[];
